Fix clippy warnings across workspace [claude-opus-4-8]
Resolve all lints flagged by clippy 1.95 under `-D warnings`. Notable
changes: box large enum variants (FrameWriter::Mp4, MaybeValidBraidzFile::Valid),
add type aliases for complex types, rename as_valid -> into_valid, add # Safety
docs to unsafe fns and C callbacks, and preserve NaN semantics in the intrinsics
positivity check. Remaining changes are mechanical (needless borrows/casts,
or_default, let-chains, etc.).

+++ im-proc/fastfreeimage/src/lib.rs
@@ -252,7 +252,7 @@ where
 #[test]
 fn test_image_view_size() -> Result<()> {
     // Create a 3x3 image with a stride of 10 bytes
-    let data = aligned_vec::avec![42; 30 as usize];
+    let data = aligned_vec::avec![42; 30_usize];
     let data = data.into_boxed_slice();
     let mut im = FastImageData::<u8> {
         data,
